refactor: encodeMessageData
